* frame.c (make_frame, x_set_frame_parameters): Use bool for boolean. (x_figure_window_size): Likewise.  Adjust to return long. (syms_of_frame): Do not DEFSYM Qterminal_live_p. (toplevel): Move Qterminal_live_p to... * terminal.c (toplevel): ...here, make it static, and... (syms_of_terminal): ...DEFSYM here. * frame.h (Qterminal_live_p): Remove declaration. (make_frame, x_figure_window_size): Adjust prototype. * nsfns.m (Fx_create_frame): Use long for window flags.

+++ src/nsfns.m
@@ -1068,7 +1068,7 @@ This function is an internal primitive--use `make-frame' instead.  */)
   Lisp_Object frame, tem;
   Lisp_Object name;
   int minibuffer_only = 0;
-  int window_prompting = 0;
+  long window_prompting = 0;
   int width, height;
   ptrdiff_t count = specpdl_ptr - specpdl;
   struct gcpro gcpro1, gcpro2, gcpro3, gcpro4;
